# backend/openapi.yaml
openapi: 3.0.3
info:
  title: CLISENSA API
  description: API REST de la plataforma multi-hospital de citas médicas CLISENSA.
  version: 0.5.0
servers:
  - url: http://localhost:4000/api
    description: Desarrollo local
...
paths:
...
  /admin/hospitales:
    post:
      tags: [Administración]
      summary: Crear hospital (ADMIN_SISTEMA)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, direccion, telefono]
              properties:
                nombre: { type: string }
                direccion: { type: string }
                telefono: { type: string }
                logoUrl: { type: string, format: uri }
      responses:
        "201": { description: Hospital creado }
        "403": { description: Rol no autorizado }
  /admin/personal:
    post:
      tags: [Administración]
      summary: Crear personal administrativo para un hospital (ADMIN_SISTEMA)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hospitalId, nombre, correo, password]
              properties:
                hospitalId: { type: integer }
                nombre: { type: string }
                correo: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        "201": { description: Cuenta administrativa creada }
        "404": { description: Hospital no encontrado }
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión y obtener un token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [correo, password]
              properties:
                correo: { type: string, format: email, example: paciente@demo.com }
                password: { type: string, example: "1234" }
      responses:
        "200":
          description: Inicio de sesión exitoso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaAutenticacion"
        "400": { description: Faltan credenciales }
        "401": { description: Credenciales inválidas }
  /auth/register:
    post:
      tags: [Auth]
      summary: Registrar paciente e iniciar sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, correo, password, cedula]
              properties:
                nombre: { type: string, example: Ana Paciente }
                correo: { type: string, format: email, example: ana@example.com }
                password: { type: string, minLength: 8, example: "ejemplo-seguro" }
                cedula: { type: string, description: Identificador requerido del paciente }
      responses:
        "201":
          description: Paciente registrado y autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaAutenticacion"
        "400": { description: Faltan campos requeridos o el correo tiene formato inválido }
        "409": { description: El correo o la cédula ya están registrados }
  /auth/perfil:
    get:
      tags: [Auth]
      summary: Consultar el perfil de la cuenta autenticada
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Datos del perfil autenticado }
        "401": { description: Falta autenticación }
    put:
      tags: [Auth]
      summary: Actualizar datos personales, foto o contraseña
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, correo]
              properties:
                nombre: { type: string }
                correo: { type: string, format: email }
                cedula: { type: string, description: Requerida para pacientes }
                fotoUrl: { type: string, format: uri, nullable: true }
                passwordActual: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        "200": { description: Perfil actualizado }
        "400": { description: Datos inválidos o contraseña actual incorrecta }
        "409": { description: El correo o la cédula ya están registrados }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UsuarioPublico:
      type: object
      required: [id, nombre, correo, rol, medicoId]
      properties:
        id: { type: integer }
        nombre: { type: string }
        correo: { type: string, format: email }
        rol: { type: string, enum: [ADMIN_SISTEMA, PERSONAL_ADMINISTRATIVO, MEDICO, PACIENTE] }
        medicoId: { type: integer, nullable: true }
        hospitalId: { type: integer, nullable: true }
        cedula: { type: string, nullable: true }
        fotoUrl: { type: string, format: uri, nullable: true }
    RespuestaAutenticacion:
      type: object
      required: [token, usuario]
      properties:
        token: { type: string }
        usuario:
          $ref: "#/components/schemas/UsuarioPublico"
...